+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 - Replace privacy policy with the terms and conditions in root screen 📄
 - Hide inquisitor sign in form 😶‍🌫️
+- Skip enhancement step in army builder with only heroes ⏭️
 
 ### Fixed